Update to conform to the verdi-click interface.
With the verdi-click update, most of the fixtures are just shallow
wrappers invoking the corresponding click command. The longer-term
goal is to get rid of them almost entirely, and rely on the fixtures
provided by AiiDA itself. This also ensures consistent interfaces.
For now, we keep the 'translation layer' to ensure backwards
compatibility.